style: Move font-synthesis outside of mako
